+++ packages/react/src/experimental/Forms/AvatarNameSelector/index.tsx
@@ -199,9 +199,8 @@ export const AvatarNameSelector = (
         }
       }
     } else {
-      const visibleIds = new Set(filteredEntities.map((ent) => ent.id))
-      newSelectedEntities = (prevSelected ?? []).filter((el) => {
-        return visibleIds.has(el.id) && el.id !== entityToRemove.id
+      newSelectedEntities = (prevSelected ?? []).filter((selectedEntity) => {
+        return selectedEntity.id !== entityToRemove.id
       })
     }