+++ packages/react/src/experimental/OneDataCollection/item-actions.tsx
@@ -1,33 +1,17 @@
 import { DropdownItem } from "../exports"
+
 import { RecordType } from "./types"
 
-import { DropdownItemObject } from "../exports"
-
-/**
- * Defines the structure and configuration of the primary action that can be performed on a collection.
- * @returns An action
- */
-export type PrimaryActionsDefinition = () =>
-  | Pick<DropdownItemObject, "onClick" | "label" | "icon">
-  | undefined
-
-/**
- * Defines the structure and configuration of secondary actions that can be performed on a collection.
- * @returns An array of actions
- */
-export type SecondaryActionsDefinition = () =>
-  | Array<DropdownItem & { enabled?: boolean }>
-  | undefined
-
-/**
- * Filters the actions based on the enabled property
- * @param actions - The actions to filter
- * @returns An array of filtered actions
- */
 export type ItemActionsDefinition<T extends RecordType> = (
   item: T
 ) => Array<DropdownItem & { enabled?: boolean }> | undefined
 
+/**
+ * Filters the actions based on the enabled property
+ * @param actions - The actions to filter
+ * @param item - The item to filter the actions for
+ * @returns An array of filtered actions
+ */
 export const filterItemActions = <T extends RecordType>(
   actions: ItemActionsDefinition<T>,
   item: T