+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/Table/index.tsx
@@ -1,3 +1,4 @@
+import { Checkbox } from "@/experimental/Forms/Fields/Checkbox"
 import { OnePagination } from "@/experimental/OnePagination"
 import {
   OneTable,
@@ -7,10 +8,9 @@ import {
   TableHeader,
   TableRow,
 } from "@/experimental/OneTable"
-import { ColumnWidth } from "@/experimental/OneTable/utils/sizes"
 import { useI18n } from "@/lib/i18n-provider"
 import { cn } from "@/lib/utils"
-import { ComponentProps } from "react"
+import { ComponentProps, useMemo } from "react"
 import type { FiltersDefinition } from "../../../Filters/types"
 import { ItemActionsDefinition } from "../../../item-actions"
 import { ActionsDropdown } from "../../../ItemActions/Dropdown"
@@ -22,6 +22,7 @@ import {
 } from "../../../sortings"
 import { CollectionProps, RecordType } from "../../../types"
 import { useData } from "../../../useData"
+import { useSelectable } from "../../../useSelectable"
 
 export type WithOptionalSorting<
   Record,
@@ -37,7 +38,7 @@ export type WithOptionalSorting<
   /**
    * The width of the column. If not provided, the width will be "auto"
    */
-  width?: ColumnWidth
+  width?: number
 }
 
 export type TableColumnDefinition<
@@ -52,6 +53,7 @@ export type TableVisualizationOptions<
   Sortings extends SortingsDefinition,
 > = {
   columns: ReadonlyArray<TableColumnDefinition<Record, Sortings>>
+  frozenColumns?: 0 | 1 | 2
 }
 
 export const TableCollection = <
@@ -62,6 +64,8 @@ export const TableCollection = <
 >({
   columns,
   source,
+  frozenColumns = 0,
+  onSelectItems,
 }: CollectionProps<
   Record,
   Filters,
@@ -78,6 +82,20 @@ export const TableCollection = <
   >(source)
 
   const { currentSortings, setCurrentSortings, isLoading } = source
+
+  const frozenColumnsLeft = useMemo(() => frozenColumns, [frozenColumns])
+
+  /**
+   * Item selection
+   */
+
+  const {
+    selectedItems,
+    isAllSelected,
+    isPartiallySelected,
+    handleSelectItemChange,
+    handleSelectAll,
+  } = useSelectable(data, paginationInfo, source, onSelectItems)
 
   /**
    * Determine the sort state of a column
@@ -147,12 +165,25 @@ export const TableCollection = <
     return renderProperty(item, column, "table")
   }
 
+  const checkColumnWidth = source.selectable ? 50 : 0
+
   return (
     <>
       <OneTable loading={isLoading}>
         <TableHeader>
           <TableRow>
-            {columns.map(({ sorting, label, ...column }) => (
+            {source.selectable && (
+              <TableHead width={checkColumnWidth} sticky={{ left: 0 }}>
+                <Checkbox
+                  checked={isAllSelected}
+                  indeterminate={isPartiallySelected}
+                  onCheckedChange={handleSelectAll}
+                  title="Select all"
+                  hideLabel
+                />
+              </TableHead>
+            )}
+            {columns.map(({ sorting, label, ...column }, index) => (
               <TableHead
                 key={String(label)}
                 sortState={getColumnSortState(
@@ -160,6 +191,19 @@ export const TableCollection = <
                   source.sortings,
                   currentSortings
                 )}
+                width={column.width}
+                sticky={
+                  index < frozenColumnsLeft
+                    ? {
+                        left: columns
+                          .slice(0, Math.max(0, index))
+                          .reduce(
+                            (acc, column) => acc + (column.width ?? 0),
+                            checkColumnWidth
+                          ),
+                      }
+                    : undefined
+                }
                 {...column}
                 onSortClick={
                   sorting
@@ -174,7 +218,14 @@ export const TableCollection = <
               </TableHead>
             ))}
             {source.itemActions && (
-              <TableHead key="actions" width="fit" hidden>
+              <TableHead
+                key="actions"
+                width={50}
+                hidden
+                sticky={{
+                  right: 0,
+                }}
+              >
                 {t.collections.actions.actions}
               </TableHead>
             )}
@@ -183,15 +234,41 @@ export const TableCollection = <
         <TableBody>
           {data.map((item, index) => {
             const itemHref = source.itemUrl ? source.itemUrl(item) : undefined
-
+            const id = source.selectable ? source.selectable(item) : undefined
             return (
               <TableRow key={`row-${index}`}>
+                {source.selectable && (
+                  <TableCell width={checkColumnWidth} sticky={{ left: 0 }}>
+                    {id !== undefined && (
+                      <Checkbox
+                        checked={selectedItems.has(id)}
+                        onCheckedChange={(checked) =>
+                          handleSelectItemChange(item, checked)
+                        }
+                        title={`Select ${source.selectable(item)}`}
+                        hideLabel
+                      />
+                    )}
+                  </TableCell>
+                )}
                 {columns.map((column, cellIndex) => (
                   <TableCell
                     key={String(column.label)}
                     firstCell={cellIndex === 0}
                     href={itemHref}
-                    sticky={column.sticky}
+                    width={column.width}
+                    sticky={
+                      cellIndex < frozenColumnsLeft
+                        ? {
+                            left: columns
+                              .slice(0, Math.max(0, cellIndex))
+                              .reduce(
+                                (acc, column) => acc + (column.width ?? 0),
+                                checkColumnWidth
+                              ),
+                          }
+                        : undefined
+                    }
                   >
                     <div
                       className={cn(
@@ -204,7 +281,14 @@ export const TableCollection = <
                   </TableCell>
                 ))}
                 {source.itemActions && (
-                  <TableCell key="actions" href={itemHref}>
+                  <TableCell
+                    key="actions"
+                    width={50}
+                    sticky={{
+                      right: 0,
+                    }}
+                    href={itemHref}
+                  >
                     <ActionsDropdown item={item} actions={source.itemActions} />
                   </TableCell>
                 )}