+++ packages/react/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.18.1](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.18.0...factorial-one-react-v1.18.1) (2025-04-01)
+
+
+### Bug Fixes
+
+* **PageHeader:** adding asChild Property ([#1516](https://github.com/factorialco/factorial-one/issues/1516)) ([e4ddcfa](https://github.com/factorialco/factorial-one/commit/e4ddcfa5cf95b841f19b56e5980f4772074141af))
+
+## [1.18.0](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.17.0...factorial-one-react-v1.18.0) (2025-03-31)
+
+
+### Features
+
+* datacolletion selectable rows ([#1489](https://github.com/factorialco/factorial-one/issues/1489)) ([e5b0ada](https://github.com/factorialco/factorial-one/commit/e5b0ada4fccda81a1c95104e3b9fb64edffe72ed))
+
+## [1.17.0](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.16.1...factorial-one-react-v1.17.0) (2025-03-31)
+
+
+### Features
+
+* datacollection / onetable frozen cols ([444ad94](https://github.com/factorialco/factorial-one/commit/444ad9446f83c201c341bbb7d26a542290e55090))
+
 ## [1.16.1](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.16.0...factorial-one-react-v1.16.1) (2025-03-31)